% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonction_graph_FY.R
\name{graph_FY}
\alias{graph_FY}
\title{Graphique de la fonction de répartition de Y}
\usage{
graph_FY(..., type = c("via", "temp", "diff", "gar"))
}
\arguments{
\item{...}{arguments supplémentaires fournis à la fonction produisant le
graphique (fonction déterminée selon l'argument « type »).}

\item{type}{chaîne de caractères représentant le type de la rente. La valeur
par défaut est "via". Les choix sont les suivants :
\itemize{
\item "via" : rente viagère
\item "temp" : rente temporaire
\item "diff" : rente différée
\item "gar" : rente garantie
}}
}
\value{
graphique.
}
\description{
Cette fonction permet de produire un graphique (ggplot) présentant la
fonction de répartition de Y selon le type de rente choisi. Il s'agit, en
quelque sorte, d'une fonction « wrapper ».
}
\details{
Pour plus de détails sur la création des graphiques, consultez les fonctions
associées à chaque type de rente.
}
