% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonction_graph_FY_via.R
\name{graph_FY_via}
\alias{graph_FY_via}
\title{Graphique de la fonction de répartition de Y (rente viagère)}
\usage{
graph_FY_via(Y, t, tqx, n, couleur = "black")
}
\arguments{
\item{Y}{vecteur de nombres réels représentant les valeurs actualisées.}

\item{t}{vecteur de nombres réels représentant les durées utilisées pour
calculer les valeurs actualisées et les probabilités de décès.}

\item{tqx}{vecteur de nombres réels représentant les probabilités de décès.}

\item{n}{argument fantôme (voir les détails).}

\item{couleur}{chaîne de caractères définissant la couleur des courbes. La
valeur par défaut est "black".}
}
\value{
graphique.
}
\description{
Cette fonction produit un graphique (ggplot) de la fonction de répartition
pour la variable aléatoire Y dans le cas d'une rente viagère.
}
\details{
Les paramètres « Y », « t » et « tqx » doivent être de la même longueur.
Normalement, les trois arguments doivent provenir du même vecteur « t ».
L'argument « n » n'est pas utilisé; sa présence simplifie le code de
l'application Shiny.
}
