% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonction_graph_Y.R
\name{graph_Y}
\alias{graph_Y}
\title{Graphique de Y en fonction de T}
\usage{
graph_Y(..., x, type = c("via", "temp", "diff", "gar"))
}
\arguments{
\item{...}{arguments supplémentaires fournis à la fonction produisant le
graphique (fonction déterminée selon l'argument « type »).}

\item{x}{nombre réel positif représentant l'âge initial de la personne
assurée.}

\item{type}{chaîne de caractères représentant le type de la rente. La valeur
par défaut est "via". Les choix sont les suivants :
\itemize{
\item "via" : rente viagère
\item "temp" : rente temporaire
\item "diff" : rente différée
\item "gar" : rente garantie
}}
}
\value{
graphique.
}
\description{
Cette fonction permet de produire un graphique (ggplot) de Y en fonction de T
selon le type de rente choisi. Il s'agit, en quelque sorte, d'une fonction
« wrapper ».
}
\details{
Pour plus de détails sur la création des graphiques, consultez les fonctions
associées à chaque type de rente.
}
