% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonction_graph_Y_diff.R
\name{graph_Y_diff}
\alias{graph_Y_diff}
\title{Graphique de Y en fonction de T (rente différée)}
\usage{
graph_Y_diff(Y, t, n, couleur = "red")
}
\arguments{
\item{Y}{vecteur de nombres réels représentant les valeurs actualisées.}

\item{t}{vecteur de nombres réels représentant les durées utilisées pour
calculer les valeurs actualisées et les probabilités de décès.}

\item{n}{nombre réel représentant la durée de la période différée.}

\item{couleur}{chaîne de caractères définissant la couleur des courbes. La
valeur par défaut est "red".}
}
\value{
graphique.
}
\description{
Cette fonction produit un graphique (ggplot) de Y en fonction de T dans le
cas d'une rente différée.
}
\details{
Les paramètres « Y », « t » et « tqx » doivent être de la même longueur.
Normalement, les trois arguments doivent provenir du même vecteur « t ».
}
