% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonction_graph_Y_gar.R
\name{graph_Y_gar}
\alias{graph_Y_gar}
\title{Graphique de Y en fonction de T (rente garantie)}
\usage{
graph_Y_gar(Y, t, n, couleur = "orange")
}
\arguments{
\item{Y}{vecteur de nombres réels représentant les valeurs actualisées.}

\item{t}{vecteur de nombres réels représentant les durées utilisées pour
calculer les valeurs actualisées et les probabilités de décès.}

\item{n}{nombre réel représentant la durée des paiements garantis.}

\item{couleur}{chaîne de caractères définissant la couleur des courbes. La
valeur par défaut est "orange".}
}
\value{
graphique.
}
\description{
Cette fonction produit un graphique (ggplot) de Y en fonction de T dans le
cas d'une rente garantie.
}
\details{
Les paramètres « Y », « t » et « tqx » doivent être de la même longueur.
Normalement, les trois arguments doivent provenir du même vecteur « t ».
}
