% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonction_graph_Y_temp.R
\name{graph_Y_temp}
\alias{graph_Y_temp}
\title{Graphique de Y en fonction de T (rente temporaire)}
\usage{
graph_Y_temp(Y, t, n, couleur = "blue")
}
\arguments{
\item{Y}{vecteur de nombres réels représentant les valeurs actualisées.}

\item{t}{vecteur de nombres réels représentant les durées utilisées pour
calculer les valeurs actualisées et les probabilités de décès.}

\item{n}{nombre réel représentant la durée de la rente (période temporaire).}

\item{couleur}{chaîne de caractères définissant la couleur des courbes. La
valeur par défaut est "blue".}
}
\value{
graphique.
}
\description{
Cette fonction produit un graphique (ggplot) de Y en fonction de T dans le
cas d'une rente temporaire.
}
\details{
Les paramètres « Y », « t » et « tqx » doivent être de la même longueur.
Normalement, les trois arguments doivent provenir du même vecteur « t ».
}
