% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonction_tqx.R
\name{tqx}
\alias{tqx}
\title{Fonction de répartition de la durée de vie complète}
\usage{
tqx(methode = c("demoivre", "expo"), ...)
}
\arguments{
\item{methode}{chaîne de caractères définissant la méthode (la loi/la
fonction) à utiliser.}

\item{...}{arguments supplémentaires fournis à la fonction utilisée pour
calculer la probabilité de décès.}
}
\value{
résultat dans un format spécifique à la fonction choisie à l'aide de
l'argument « methode ».
}
\description{
Cette fonction permet de calculer la probabilité de décéder dans un
intervalle de temps donné selon une loi de mortalité quelconque.
}
\details{
Les lois pouvant être utilisées actuellement sont la loi de De Moivre (voir
la fonction « tqx_demoivre ») et la loi exponentielle (voir la fonction
« tqx_expo »). Il est simple d'ajouter d'autres lois.
}
