% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonction_tqx_demoivre.R
\name{tqx_demoivre}
\alias{tqx_demoivre}
\title{Fonction de répartition de la durée de vie complète (De Moivre)}
\usage{
tqx_demoivre(x, t, omega)
}
\arguments{
\item{x}{nombre réel positif représentant l'âge de la personne.}

\item{t}{nombre réel positif représentant la durée de vie maximale (ou, de
manière équivalente, l'intervalle de temps à l'intérieur duquel la personne
va décéder). t peut aussi être un vecteur contenant plusieurs nombres.}

\item{omega}{nombre réel positif représentant l'âge ultime.}
}
\value{
nombre réel entre 0 et 1 représentant la probabilité que la personne
décède entre l'âge x et l'âge x + t. Si plusieurs t sont utilisés, la
fonction retourne un vecteur.
}
\description{
Cette fonction permet de calculer la probabilité de décéder dans un
intervalle de temps donné selon la loi de mortalité de De Moivre.
}
\details{
La loi de mortalité de De Moivre est utilisée; cette loi suppose une
distribution uniforme des décès entre l'âge initial et l'âge ultime.
}
