% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonction_tqx_expo.R
\name{tqx_expo}
\alias{tqx_expo}
\title{Fonction de répartition de la durée de vie complète (exponentielle)}
\usage{
tqx_expo(t, mu)
}
\arguments{
\item{t}{nombre réel positif représentant la durée de vie maximale (ou, de
manière équivalente, l'intervalle de temps à l'intérieur duquel la personne
va décéder). t peut aussi être un vecteur contenant plusieurs nombres.}

\item{mu}{nombre réel positif représentant la force de mortalité.}
}
\value{
nombre réel entre 0 et 1 représentant la probabilité que la personne
décède entre l'âge x et l'âge x + t. Si plusieurs t sont utilisés, la
fonction retourne un vecteur.
}
\description{
Cette fonction permet de calculer la probabilité de décéder dans un
intervalle de temps donné selon la loi exponentielle.
}
\details{
La loi exponentielle est utilisée; puisque cette loi est sans mémoire, l'âge
initial de la personne n'a pas d'importance.
}
