% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonction_Z.R
\name{Z}
\alias{Z}
\title{Fonction pour calculer la valeur actualisée d'une prestation}
\usage{
Z(delta, t)
}
\arguments{
\item{delta}{nombre réel positif représentant la force d'intérêt.}

\item{t}{nombre réel positif représentant la durée de vie (ou, de manière
équivalente, le temps avant le décès).}
}
\value{
nombre réel entre 0 et 1 représentant la valeur actualisée.
}
\description{
Cette fonction permet de calculer la valeur actualisée d'une prestation de 1$
en fonction de la force d'intérêt et de la durée de vie.
}
\details{
La fonction pourrait être complexifiée en permettant l'utilisation d'une
force d'intérêt qui varie dans le temps.
}
